# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LuauHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScriptStatus = typeof ScriptStatus[keyof typeof ScriptStatus];


export const ScriptStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ScriptObfuscationStatus = typeof ScriptObfuscationStatus[keyof typeof ScriptObfuscationStatus];


export const ScriptObfuscationStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface Script {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  obfuscatedCode?: string | null;
  /** @nullable */
  scriptKey?: string | null;
  /** @nullable */
  service?: string | null;
  status: ScriptStatus;
  obfuscationStatus: ScriptObfuscationStatus;
  executions: number;
  successCount: number;
  failureCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ScriptInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  code?: string;
  service?: string;
}

export type ScriptUpdateStatus = typeof ScriptUpdateStatus[keyof typeof ScriptUpdateStatus];


export const ScriptUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface ScriptUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  code?: string;
  service?: string;
  status?: ScriptUpdateStatus;
}

export type ExecutionLogStatus = typeof ExecutionLogStatus[keyof typeof ExecutionLogStatus];


export const ExecutionLogStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export interface ExecutionLog {
  id: number;
  scriptId: number;
  scriptName: string;
  status: ExecutionLogStatus;
  executorId: string;
  /** @nullable */
  robloxUsername?: string | null;
  /** @nullable */
  robloxUserId?: string | null;
  /** @nullable */
  robloxClientId?: string | null;
  /** @nullable */
  robloxThumbnailUrl?: string | null;
  /** @nullable */
  robloxExecutor?: string | null;
  /** @nullable */
  duration?: number | null;
  createdAt: string;
}

export type ExecutionLogInputStatus = typeof ExecutionLogInputStatus[keyof typeof ExecutionLogInputStatus];


export const ExecutionLogInputStatus = {
  success: 'success',
  failed: 'failed',
} as const;

export interface ExecutionLogInput {
  scriptId: number;
  status: ExecutionLogInputStatus;
  executorId: string;
  duration?: number;
}

export interface AnalyticsOverview {
  totalScripts: number;
  totalExecutions: number;
  activeUsers: number;
  successRate: number;
  activeScripts: number;
  obfuscatedScripts: number;
}

export interface ChartDataPoint {
  date: string;
  executions: number;
  success: number;
  failed: number;
}

export type ListScriptsParams = {
search?: string;
status?: ListScriptsStatus;
};

export type ListScriptsStatus = typeof ListScriptsStatus[keyof typeof ListScriptsStatus];


export const ListScriptsStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type ListLogsParams = {
status?: ListLogsStatus;
limit?: number;
};

export type ListLogsStatus = typeof ListLogsStatus[keyof typeof ListLogsStatus];


export const ListLogsStatus = {
  success: 'success',
  failed: 'failed',
} as const;